Seriously improve fish location via thresholding
This commit marks major improvements to the background
subtraction/thresholding operation that is used to locate the fish. The
class is now called LocateFish, as it's specialized to its purpose.

Summary
  - Thresholded image contains only fish body (most of the time)
  - Order of outputs changed
  - Some error checking added
  - Code heavily commented

Morphological operations (opening, dilating) are used to remove other
debris the binary image, as well as capture the entirety of the fish's
body. Connected components operations are successful at ignoring
shadows, except when the fish is at the bottom of the tank.