Replace the old ShowFeatures class
Make slight modifications to the FishOrientation class so it will
replace the functionality of ShowFeatures when only one feature is
supplied. That is, its normal purpose is to display two features, and
make calculations based off the distance between them. With just one
feature, all it can do is draw the feature.

Accordingly, all instances of ShowFeatures are replaced with the
FishOrientation class (eventually, they'll be tracking two features
anyway, so this is a sensible change).